demos/adk-with-skills/banana_ric_agent_tmp20260429_145522/requirements.txt
google-genai
Pillow
google-adk
google-cloud-aiplatform
google-cloud-storage
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-grpc
opentelemetry-instrumentation-httpx
opentelemetry-instrumentation-google-genai
